DRY test case
